Enable version constraints in AutoDependOn tag
Make it possible to specify packages with versions as the value of
AutoDependOn tag inside Package Information File. The syntax is the same
as for the Dependencies tag.

This patch is not in a shape suitable for upstream inclusion. It changes
an API unnecessary and the proper solution should be to honor version
constrains from Dependencies tag instead.

Task-number: QTIFW-1165
Change-Id: Ifc7f54e5ccd1bd9f82de0e4596882de05611a92b