Updater: uninstall old components in correct order
If a dependency is removed from a new version of a component, the
updater will uninstall the old components in a random order, ignoring
the dependency existing between the old components.

This is because the stored undo operations are reordered after the
maintenance tool is updated, using the new metadata, where the
dependency does not exist anymore.

This commit fixes this by sorting based on the metadata of the installed
packages, instead of the packages available from online repositories.

With the original approach of sorting operations prior to writing the
configuration, this change would only become effective with the next
maintenance tool update. In order to make the change effective
immediately, operations are reordered unconditionally after loading the
configuration since now.

Task-number: QTIFW-1166
Change-Id: I95abcb6530aae5805b9135d6ee32944ce5095592